requirements.txt
Django
psycopg2-binary
django-crispy-forms
djangorestframework
django-bootstrap3